requirements.txt
google-generativeai>=0.3.0
selenium>=4.0.0
python-dotenv>=1.0.0
